refactor Ray's worker to accept ConfigDSI. STABLE
