--- dsi/offline/heatmap/ray_manager.py
@@ -1,56 +1,0 @@
-import logging
-
-import pandas as pd
-import ray
-from ray.experimental import tqdm_ray
-
-from dsi.configs.experiment.simul.heatmap import ConfigHeatmap
-from dsi.offline.heatmap.params import get_df_heatmap_params
-from dsi.offline.heatmap.worker import get_all_latencies
-from dsi.types.name import Param
-
-log = logging.getLogger(__name__)
-
-
-class RayManager:
-    def __init__(self, config: ConfigHeatmap) -> None:
-        self._df_configs: pd.DataFrame = get_df_heatmap_params(config)
-        self._results_raw: None | list[tuple[int, dict[str, float]]] = None
-        self.df_results: pd.DataFrame = self._df_configs.copy(deep=True)
-
-    def run(self) -> pd.DataFrame:
-        # NOTE: Ray discovers and utilizes all available resources by default
-        ray.init(
-            ignore_reinit_error=True,
-        )
-        remote_tqdm = ray.remote(tqdm_ray.tqdm)
-        bar = remote_tqdm.remote(total=len(self._df_configs))
-        futures = [
-            self._process_row.remote(index=index, row=row, bar=bar)
-            for index, row in self._df_configs.iterrows()
-        ]
-        self._results_raw = ray.get(futures)
-        bar.close.remote()
-        ray.shutdown()
-        self._merge_results()
-        return self.df_results
-
-    @staticmethod
-    @ray.remote
-    def _process_row(
-        index: int, row: dict, bar: tqdm_ray.tqdm
-    ) -> tuple[int, dict[str, float]]:
-        c: float = row[Param.c]
-        a: float = row[Param.a]
-        k: int = int(row[Param.k])
-        num_target_servers: int = int(row[Param.num_target_servers])
-        all_latencies: dict[str, float] = get_all_latencies(
-            c=c, a=a, k=k, num_target_servers=num_target_servers
-        )
-        bar.update.remote(1)
-        return index, all_latencies
-
-    def _merge_results(self) -> None:
-        for i, res in self._results_raw:
-            for key, val in res.items():
-                self.df_results.at[i, key] = val
